requirements.txt
torch
torchvision
tqdm
tensorboardx
torchsummary
numpy
pillow